+++ core/src/commonTest/kotlin/com/deezer/caupain/DependencyUpdateCheckerTest.kt
@@ -65,7 +65,7 @@ import io.ktor.http.appendPathSegments
 import io.ktor.http.headersOf
 import io.ktor.http.takeFrom
 import io.ktor.serialization.kotlinx.json.json
-import io.ktor.serialization.kotlinx.xml.xml
+import io.ktor.serialization.kotlinx.serialization
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
@@ -127,7 +127,7 @@ class DependencyUpdateCheckerTest {
             httpClient = HttpClient(engine) {
                 install(ContentNegotiation) {
                     json(DefaultJson)
-                    xml(DefaultXml, ContentType.Any)
+                    serialization(ContentType.Any, DefaultXml)
                 }
             },
             ioDispatcher = testDispatcher,
@@ -268,7 +268,7 @@ class DependencyUpdateCheckerTest {
             httpClient = HttpClient(engine) {
                 install(ContentNegotiation) {
                     json(DefaultJson)
-                    xml(DefaultXml, ContentType.Any)
+                    serialization(ContentType.Any, DefaultXml)
                 }
             },
             ioDispatcher = testDispatcher,