+++ CHANGELOG.md
@@ -14,6 +14,13 @@
 
 ### Security
 
+## 1.8.1 - 2026-01-19
+
+### Fixed
+
+- Properly create directories for output files
+- Update XML parser to avoid some parsing issues for specific POM files
+
 ## 1.8.0 - 2026-01-13
 
 ### Added